P27-vault-features/assetarc-vault-features/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.1
python-dotenv==1.0.1
boto3==1.34.158
Pillow==10.3.0
